fix sp_async
