+++ llm/utils/fused_layers.py
@@ -18,11 +18,15 @@ import paddle.distributed as dist
 from paddle import _C_ops
 from paddle.distributed import fleet
 from paddle.distributed.fleet.utils.log_util import logger
+from paddle.distributed.fleet.utils.sequence_parallel_utils import (
+    _check_environment_for_overlap,
+)
 from paddle.framework import core
 
 _raise_cuda_env_unset_warning = True
 _mp_async_allreduce = False
 _sp_async_reduce_scatter = False
+paddle_nn_functional_linear = paddle.nn.functional.linear
 
 
 def is_fused_matmul_bias_supported():
@@ -34,6 +38,7 @@ def is_fused_matmul_bias_supported():
 
 if is_fused_matmul_bias_supported():
     origin_linear = paddle.incubate.nn.functional.fused_linear
+    paddle_incubate_nn_functional_fused_linear = paddle.incubate.nn.functional.fused_linear
 else:
     origin_linear = paddle.nn.functional.linear
 
@@ -87,6 +92,38 @@ def sync_mp_allreduce(task, dist_tensor):
         placements=new_placments,
         place=place,
     )
+
+
+# modify from Paddle/python/paddle/distributed/auto_parallel/moe_utils.py
+def _dist_reshape(
+    dist_tensor,
+    global_shape,
+    mesh,
+    placements,
+):
+    local_tensor = dist_tensor._local_value()
+    tgt_global_shape = [dist_tensor.shape[0] * dist_tensor.shape[1], dist_tensor.shape[2]]
+    tgt_local_shape = [local_tensor.shape[0] * local_tensor.shape[1], local_tensor.shape[2]]
+
+    place = paddle.framework._current_expected_place()
+    place = paddle.framework._get_paddle_place(place)
+
+    local_tensor = local_tensor.reshape(tgt_local_shape)
+
+    if placements[1].is_shard():
+        new_placements = [dist.Shard(0), dist.Shard(1)]
+    else:
+        new_placements = [dist.Shard(0), dist.Replicate()]
+
+    out = paddle.Tensor(
+        local_tensor,
+        dims=tgt_global_shape,
+        process_mesh=mesh,
+        placements=new_placements,
+        place=place,
+    )
+    out.stop_gradient = dist_tensor.stop_gradient
+    return out
 
 
 class FusedLinearWithGradAdd(paddle.autograd.PyLayer):
@@ -180,8 +217,196 @@ class OverlapLinear(paddle.autograd.PyLayer):
             return x_grad, weight_grad, bias_grad
 
 
+class FusedLinearWithReduceScatter(paddle.autograd.PyLayer):
+    @staticmethod
+    def forward(ctx, x, weight, bias=None, name=None):
+        input_parallel = dist.reshard(
+            x,
+            x.process_mesh,
+            [dist.Shard(1), dist.Replicate()],
+        )
+        y = origin_linear(input_parallel, weight, bias)
+        ctx.save_for_backward(weight, bias, input_parallel)
+
+        return y
+
+    @staticmethod
+    def backward(ctx, dy):
+        weight, bias, input_parallel = ctx.saved_tensor()
+
+        # compute dx
+        if dy.dtype == weight.dtype:
+            dinput_parallel = paddle.matmul(dy, weight, transpose_y=True)
+        else:
+            dinput_parallel = paddle.matmul(dy, paddle.cast(weight, dtype=dy.dtype), transpose_y=True)
+
+        hcg = fleet.get_hybrid_communicate_group()
+        model_parallel_group = hcg.get_model_parallel_group()
+        parallelism = model_parallel_group.nranks
+
+        assert (
+            dinput_parallel.shape[0] % parallelism == 0
+        ), f"Input sequence length {dinput_parallel.shape[0]} can't be divided exactly by sequence parallelism {parallelism}"
+
+        # reduce-scatter dx
+        dx_global_shape = dinput_parallel.shape
+        dx_global_shape[0] = dx_global_shape[0] // parallelism
+        dinput_parallel_local = dinput_parallel._local_value()
+        dx_local_shape = dinput_parallel_local.shape
+        dx_local_shape[0] = dx_local_shape[0] // parallelism
+        dx_local = paddle.empty(shape=dx_local_shape, dtype=dinput_parallel.dtype)
+        task = dist.stream.reduce_scatter(
+            dx_local,
+            dinput_parallel_local,
+            op=dist.ReduceOp.SUM,
+            group=model_parallel_group,
+            sync_op=False,
+        )
+
+        # compute dw and dbias
+        _check_environment_for_overlap()
+
+        place = paddle.framework._current_expected_place()
+        place = paddle.framework._get_paddle_place(place)
+
+        if _sp_async_reduce_scatter:
+            if bias is None:
+                if hasattr(weight, "main_grad"):
+                    (weight.main_grad, _,) = paddle._C_ops.fused_linear_param_grad_add(
+                        input_parallel, dy, weight.main_grad, None, True, False
+                    )
+                    task.wait()
+                    dx = paddle.Tensor(
+                        dx_local,
+                        dims=dx_global_shape,
+                        process_mesh=dinput_parallel.process_mesh,
+                        placements=[dist.Shard(1), dist.Shard(0)],
+                        place=place,
+                    )
+                    dx.stop_gradient = dx.stop_gradient
+                    return dx, None
+                else:
+                    if weight.grad is not None:
+                        (weight.grad, _,) = paddle._C_ops.fused_linear_param_grad_add(
+                            input_parallel, dy, weight.grad, None, False, False
+                        )
+                        task.wait()
+                        dx = paddle.Tensor(
+                            dx_local,
+                            dims=dx_global_shape,
+                            process_mesh=dinput_parallel.process_mesh,
+                            placements=[dist.Shard(1), dist.Shard(0)],
+                            place=place,
+                        )
+                        dx.stop_gradient = dx.stop_gradient
+                        return dx, None
+                    else:
+                        (
+                            dw,
+                            _,
+                        ) = paddle._C_ops.fused_linear_param_grad_add(input_parallel, dy, None, None, False, False)
+                        task.wait()
+                        dx = paddle.Tensor(
+                            dx_local,
+                            dims=dx_global_shape,
+                            process_mesh=dinput_parallel.process_mesh,
+                            placements=[dist.Shard(1), dist.Shard(0)],
+                            place=place,
+                        )
+                        dx.stop_gradient = dx.stop_gradient
+                        return dx, dw
+
+            if hasattr(weight, "main_grad") and hasattr(bias, "main_grad"):
+                (weight.main_grad, bias.main_grad,) = paddle._C_ops.fused_linear_param_grad_add(
+                    input_parallel,
+                    dy,
+                    weight.main_grad,
+                    bias.main_grad,
+                    True,
+                    True,
+                )
+                task.wait()
+                dx = paddle.Tensor(
+                    dx_local,
+                    dims=dx_global_shape,
+                    process_mesh=dinput_parallel.process_mesh,
+                    placements=[dist.Shard(1), dist.Shard(0)],
+                    place=place,
+                )
+                dx.stop_gradient = dx.stop_gradient
+                return dx, None, None
+            else:
+                if weight.grad is not None:
+                    assert bias.grad is not None
+                    (weight.grad, bias.grad,) = paddle._C_ops.fused_linear_param_grad_add(
+                        input_parallel, dy, weight.grad, bias.grad, False, True
+                    )
+                    task.wait()
+                    dx = paddle.Tensor(
+                        dx_local,
+                        dims=dx_global_shape,
+                        process_mesh=dinput_parallel.process_mesh,
+                        placements=[dist.Shard(1), dist.Shard(0)],
+                        place=place,
+                    )
+                    dx.stop_gradient = dx.stop_gradient
+                    return dx, None, None
+                else:
+                    # When main_grad is not enabled and gradient_accumulation is used, the grad is not initialized for the first acc step.
+                    (
+                        dw,
+                        dbias,
+                    ) = paddle._C_ops.fused_linear_param_grad_add(input_parallel, dy, None, None, False, True)
+                    task.wait()
+                    dx = paddle.Tensor(
+                        dx_local,
+                        dims=dx_global_shape,
+                        process_mesh=dinput_parallel.process_mesh,
+                        placements=[dist.Shard(1), dist.Shard(0)],
+                        place=place,
+                    )
+                    dx.stop_gradient = dx.stop_gradient
+                    return dx, dw, dbias
+        else:
+            dy = _dist_reshape(dy, [-1, dy.shape[-1]], dy.process_mesh, dy.placements)
+            input_parallel = _dist_reshape(
+                input_parallel, [-1, input_parallel.shape[-1]], input_parallel.process_mesh, input_parallel.placements
+            )
+            dw = paddle.matmul(
+                input_parallel,
+                dy,
+                transpose_x=True,
+            )
+            if bias is None:
+                task.wait()
+                dx = paddle.Tensor(
+                    dx_local,
+                    dims=dx_global_shape,
+                    process_mesh=dinput_parallel.process_mesh,
+                    placements=[dist.Shard(1), dist.Shard(0)],
+                    place=place,
+                )
+                dx.stop_gradient = dx.stop_gradient
+                return dx, dw
+            else:
+                dbias = paddle.sum(dy, axis=0)
+                task.wait()
+                dx = paddle.Tensor(
+                    dx_local,
+                    dims=dx_global_shape,
+                    process_mesh=dinput_parallel.process_mesh,
+                    placements=[dist.Shard(1), dist.Shard(0)],
+                    place=place,
+                )
+                dx.stop_gradient = dx.stop_gradient
+                return dx, dw, dbias
+
+
 def mock_layers(
-    enable_fused_linear_grad_add=True, enable_mp_async_allreduce=False, enable_sp_async_reduce_scatter=False
+    model=None,
+    enable_fused_linear_grad_add=True,
+    enable_mp_async_allreduce=False,
+    enable_sp_async_reduce_scatter=False,
 ):
     global _mp_async_allreduce
     global _sp_async_reduce_scatter
@@ -196,3 +421,21 @@ def mock_layers(
         paddle.nn.functional.linear = OverlapLinear.apply
         if is_fused_matmul_bias_supported():
             paddle.incubate.nn.functional.fused_linear = OverlapLinear.apply
+
+    if enable_sp_async_reduce_scatter:
+
+        def forward_pre_hook(layer, input):
+            paddle.nn.functional.linear = FusedLinearWithReduceScatter.apply
+            if is_fused_matmul_bias_supported():
+                paddle.incubate.nn.functional.fused_linear = FusedLinearWithReduceScatter.apply
+
+        def forward_post_hook(layer, input, ouput):
+            paddle.nn.functional.linear = paddle_nn_functional_linear
+            if is_fused_matmul_bias_supported():
+                paddle.incubate.nn.functional.fused_linear = paddle_incubate_nn_functional_fused_linear
+
+        for name, layer in model.named_sublayers():
+            for n in ["qkv_proj", "q_proj", "k_proj", "v_proj", "gate_up_fused_proj", "gate_proj", "up_proj"]:
+                if name.endswith(n):
+                    layer.register_forward_pre_hook(forward_pre_hook)
+                    layer.register_forward_post_hook(forward_post_hook)
